Refactor inner class IOSUiViewController to its own class
