+++ backends/gdx-backend-robovm/src/com/badlogic/gdx/backends/iosrobovm/IOSApplication.java
@@ -164,8 +164,8 @@ public class IOSApplication implements Application {
 		 return new IOSGraphics(this, config, input, config.useGL30);
 	}
 
-	protected IOSGraphics.IOSUIViewController createUIViewController (IOSGraphics graphics) {
-		return new IOSGraphics.IOSUIViewController(this, graphics);
+	protected IOSUIViewController createUIViewController (IOSGraphics graphics) {
+		return new IOSUIViewController(this, graphics);
 	}
 
 	protected IOSInput createInput() {
